pyproject.toml
[project]
name = "fakabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "pydantic>=2.10.6",
    "pydantic-core>=2.27.2",
    "pydantic-settings>=2.8.1",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.1",
    "python-youtube>=0.9.7",
    "yt-dlp>=2025.2.19",
]
